+++ crates/module-system/sov-modules-macros/src/dispatch/genesis.rs
@@ -1,7 +1,9 @@
 use proc_macro2::Span;
 use syn::{DeriveInput, ImplGenerics, TypeGenerics, WhereClause};
 
-use crate::common::{get_generics_type_param, StructFieldExtractor, StructNamedField};
+use crate::common::{
+    get_derived_enum_attrs, get_generics_type_param, StructFieldExtractor, StructNamedField,
+};
 
 pub(crate) struct GenesisMacro {
     field_extractor: StructFieldExtractor,
@@ -18,6 +20,11 @@ impl GenesisMacro {
         &self,
         input: DeriveInput,
     ) -> syn::Result<proc_macro::TokenStream> {
+        let default_attrs = vec![quote::quote! {
+            #[derive(Clone, ::serde::Deserialize, ::serde::Serialize)]
+        }];
+        let config_attributes = get_derived_enum_attrs("genesis", &input, default_attrs)?;
+
         let DeriveInput {
             data,
             ident,
@@ -29,8 +36,13 @@ impl GenesisMacro {
 
         let fields = self.field_extractor.get_fields_from_struct(&data)?;
         let generic_param = get_generics_type_param(&generics, Span::call_site())?;
-        let genesis_config =
-            Self::make_genesis_config(&fields, &impl_generics, &type_generics, where_clause);
+        let genesis_config = Self::make_genesis_config(
+            &fields,
+            &impl_generics,
+            &type_generics,
+            where_clause,
+            &config_attributes,
+        );
         let genesis_fn_body = Self::make_genesis_fn_body(&fields);
 
         // Implements the Genesis trait
@@ -88,6 +100,7 @@ impl GenesisMacro {
         impl_generics: &ImplGenerics,
         type_generics: &TypeGenerics,
         where_clause: Option<&WhereClause>,
+        attributes: &[proc_macro2::TokenStream],
     ) -> proc_macro2::TokenStream {
         let field_names = fields.iter().map(|field| &field.ident);
 
@@ -105,7 +118,7 @@ impl GenesisMacro {
 
         quote::quote! {
             #[doc = "Initial configuration for the rollup."]
-            #[derive(Clone, ::serde::Deserialize, ::serde::Serialize)]
+            #(#attributes)*
             pub struct GenesisConfig #impl_generics #where_clause{
                 #(#[doc = "Module configuration"] pub #fields)*
             }